// src/app/core/models/masters/bpm/process-operation-modes.ts
export interface ProcessOperationModes{
    id:number;
    title:string;
    description: string;
    status_id :number;
}
export interface ProcessOperationModesPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: ProcessOperationModes[];
}